rename index.ts to _app.ts
