+++ cli/template/addons/trpc/api-handler.ts
@@ -1,6 +1,6 @@
 // src/pages/api/trpc/[trpc].ts
 import { createNextApiHandler } from "@trpc/server/adapters/next";
-import { appRouter } from "../../../server/trpc/router";
+import { appRouter } from "../../../server/trpc/router/_app";
 import { createContext } from "../../../server/trpc/context";
 import { env } from "../../../env/server.mjs";
