+++ cli/template/addons/trpc/auth-app-router.ts
@@ -1,10 +1,11 @@
-// src/server/router/index.ts
+// src/server/trpc/router/_app.ts
 import { router } from "../trpc";
-
 import { exampleRouter } from "./example";
+import { authRouter } from "./auth";
 
 export const appRouter = router({
   example: exampleRouter,
+  auth: authRouter,
 });
 
 // export type definition of API
